+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ## [Released]
 
+## [0.3.0] 2025-11-25
+- Add Finnish language support
+- Add French language support
+- Add Polish language support
+- Add Swedish language support
+- Add Turkish language support
+- Add Emoticon sentiment analysis support
+- Update documentation with new languages and examples
+
 ## [0.2.1] 2023-10-23
 - Add Danish